Create main.py
Main python file.
This file is responsible for starting the camera, capturing image with random string as file name, calling the trained ML model (tensorflow lite) and predicting the type of waste using the captured image. Finally, the servo motor is tilted to 0 degrees if non-biodegradable, or 120 degrees if biodegradable. Initial servo angle is 45 degrees